backend/requirements.txt

fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.13.0
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.38.1
PyPDF2>=3.0.1
python-docx>=1.1.0
jira>=3.5.2
httpx>=0.25.2
pydantic-settings>=1.2.1

